+++ lib/RobotSensors/RobotSensors.h
@@ -3,8 +3,8 @@
 // Class for the ultrasound
 class Ultrasound {
 public:
-	int US_pinTrig;
-	int US_pinEcho;
+	byte US_pinTrig;
+	byte US_pinEcho;
 	float GetDistance();
 
 };
@@ -12,7 +12,7 @@ public:
 // Class for the IR sensors A02
 class IR_A02 {
 public:
-	int IR_A02pin;
+	byte IR_A02pin;
 
 	float GetDistance();
 
@@ -21,7 +21,7 @@ public:
 // Class for the IR sensors A21
 class IR_A21 {
 public:
-	int IR_A21pin;
+	byte IR_A21pin;
 
 	float GetDistance();
 
@@ -41,5 +41,5 @@ public:
 	Ultrasound Ultrasound;
 	IR_A21 A21;
 	IR_A02 A02;
-	void SetPins(int US_pinTrig, int US_pinEcho, int IR_A21pin, int IR_A02pin);
+	void SetPins(byte US_pinTrig, byte US_pinEcho, byte IR_A21pin, byte IR_A02pin);
 };
